backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
sqlalchemy
psycopg2-binary
pytest
httpx
firecrawl-py
google-generativeai
